Added claim support for the implementation
