+++ components/org.wso2.carbon.identity.sts.passive/src/main/java/org/wso2/carbon/identity/sts/passive/custom/handler/CustomClaimsHandler.java
@@ -17,19 +17,14 @@ package org.wso2.carbon.identity.sts.passive.custom.handler;
 
 import org.apache.cxf.rt.security.claims.Claim;
 import org.apache.cxf.rt.security.claims.ClaimCollection;
-import org.apache.cxf.sts.claims.ClaimTypes;
 import org.apache.cxf.sts.claims.ClaimsHandler;
 import org.apache.cxf.sts.claims.ClaimsParameters;
 import org.apache.cxf.sts.claims.ProcessedClaim;
 import org.apache.cxf.sts.claims.ProcessedClaimCollection;
-import org.opensaml.core.xml.XMLObjectBuilder;
-import org.opensaml.core.xml.XMLObjectBuilderFactory;
-import org.opensaml.core.xml.config.XMLObjectProviderRegistrySupport;
-import org.opensaml.core.xml.schema.XSInteger;
-import org.opensaml.saml.saml2.core.AttributeValue;
 
 import java.security.Principal;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
 /**
@@ -37,26 +32,16 @@ import java.util.List;
  */
 public class CustomClaimsHandler implements ClaimsHandler {
 
-    public static final String ROLE_CLAIM =
-            "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/role";
     private static List<String> knownURIs = new ArrayList<>();
+    private HashMap<String, String> claimsKeyValuePair = new HashMap<>();
 
-    static {
-        knownURIs.add(ClaimTypes.FIRSTNAME.toString());
-        knownURIs.add(ClaimTypes.LASTNAME.toString());
-        knownURIs.add(ClaimTypes.EMAILADDRESS.toString());
-        knownURIs.add(ClaimTypes.STREETADDRESS.toString());
-        knownURIs.add(ClaimTypes.MOBILEPHONE.toString());
-        knownURIs.add(ROLE_CLAIM);
-    }
-
-    private String role = "DUMMY";
-
-    public List<String> getSupportedClaimTypes() {
-
-        return knownURIs;
-    }
-
+    /**
+     * Create a processed claim collection using the claim values and params provided.
+     *
+     * @param claims The unprocessed claims.
+     * @param parameters The claim parameters.
+     * @return The processed claims.
+     */
     public ProcessedClaimCollection retrieveClaimValues(
             ClaimCollection claims, ClaimsParameters parameters) {
 
@@ -65,49 +50,9 @@ public class CustomClaimsHandler implements ClaimsHandler {
             for (Claim requestClaim : claims) {
                 ProcessedClaim claim = new ProcessedClaim();
                 claim.setClaimType(requestClaim.getClaimType());
-                if (ClaimTypes.FIRSTNAME.toString().equals(requestClaim.getClaimType())) {
-                    if (requestClaim instanceof CustomRequestClaim) {
-                        CustomRequestClaim customClaim = (CustomRequestClaim) requestClaim;
-                        String customName = customClaim.getValues().get(0) + "@"
-                                + customClaim.getScope();
-                        claim.addValue(customName);
-                    } else {
-                        claim.addValue("alice");
-                    }
-                } else if (ClaimTypes.LASTNAME.toString().equals(requestClaim.getClaimType())) {
-                    claim.addValue("doe");
-                } else if (ClaimTypes.EMAILADDRESS.toString().equals(requestClaim.getClaimType())) {
-                    claim.addValue("alice@cxf.apache.org");
-                } else if (ClaimTypes.STREETADDRESS.toString().equals(requestClaim.getClaimType())) {
-                    claim.addValue("1234 1st Street");
-                } else if (ClaimTypes.MOBILEPHONE.toString().equals(requestClaim.getClaimType())) {
-                    // Test custom (Integer) attribute value
-                    XMLObjectBuilderFactory builderFactory =
-                            XMLObjectProviderRegistrySupport.getBuilderFactory();
-
-                    @SuppressWarnings("unchecked")
-                    XMLObjectBuilder<XSInteger> xsIntegerBuilder =
-                            (XMLObjectBuilder<XSInteger>) builderFactory.getBuilder(XSInteger.TYPE_NAME);
-                    XSInteger attributeValue =
-                            xsIntegerBuilder.buildObject(AttributeValue.DEFAULT_ELEMENT_NAME, XSInteger.TYPE_NAME);
-                    attributeValue.setValue(185912592);
-
-                    claim.addValue(attributeValue);
-
-                } else if (ROLE_CLAIM.equals(requestClaim.getClaimType())) {
-                    if (requestClaim.getValues().size() > 0) {
-                        for (Object requestedRole : requestClaim.getValues()) {
-                            if (isUserInRole(parameters.getPrincipal(), requestedRole.toString())) {
-                                claim.addValue(requestedRole);
-                            }
-                        }
-                        if (claim.getValues().isEmpty()) {
-                            continue;
-                        }
-                    } else {
-                        // If no specific role was requested return DUMMY role for user
-                        claim.addValue(role);
-                    }
+                if (knownURIs.contains(requestClaim.getClaimType()) &&
+                        claimsKeyValuePair.containsKey(requestClaim.getClaimType())) {
+                    claim.addValue(claimsKeyValuePair.get(requestClaim.getClaimType()));
                 }
                 claimCollection.add(claim);
             }
@@ -122,8 +67,43 @@ public class CustomClaimsHandler implements ClaimsHandler {
         return true;
     }
 
-    public void setRole(String role) {
+    /**
+     * Get the list of supported claim URIs.
+     *
+     * @return List of supported claim URIs.
+     */
+    public List<String> getSupportedClaimTypes() {
 
-        this.role = role;
+        return knownURIs;
+    }
+
+    /**
+     * Set the list of supported claim URIs.
+     *
+     * @param knownURIs New list to be set as the known URIs.
+     */
+    public static void setKnownURIs(List<String> knownURIs) {
+
+        CustomClaimsHandler.knownURIs = knownURIs;
+    }
+
+    /**
+     * Get claim URIs and values in the form of a HashMap.
+     *
+     * @return HashMap containing the claim URIs and values.
+     */
+    public HashMap<String, String> getClaimsKeyValuePair() {
+
+        return claimsKeyValuePair;
+    }
+
+    /**
+     * Set claim key value pair(the URI and value).
+     *
+     * @param claimsKeyValuePair The new HashMap of claims key value pair.
+     */
+    public void setClaimsKeyValuePair(HashMap<String, String> claimsKeyValuePair) {
+
+        this.claimsKeyValuePair = claimsKeyValuePair;
     }
 }
