--- components/org.wso2.carbon.identity.sts.passive/src/main/java/org/wso2/carbon/identity/sts/passive/custom/handler/CustomRequestClaim.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wso2.carbon.identity.sts.passive.custom.handler;
-
-import org.apache.cxf.rt.security.claims.Claim;
-
-/**
- * Extends RequestClaim class to add additional attributes.
- */
-public class CustomRequestClaim extends Claim {
-
-    private static final long serialVersionUID = 7407723714936495457L;
-    private String scope;
-
-    public String getScope() {
-
-        return scope;
-    }
-
-    public void setScope(String scope) {
-
-        this.scope = scope;
-    }
-}
