+++ components/org.wso2.carbon.identity.sts.passive/src/main/java/org/wso2/carbon/identity/sts/passive/processors/SigningRequestProcessor.java
@@ -21,12 +21,11 @@ import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.cxf.sts.QNameConstants;
-import org.apache.cxf.sts.claims.ClaimsHandler;
 import org.apache.cxf.sts.claims.ClaimsManager;
 import org.apache.cxf.sts.operation.TokenIssueOperation;
 import org.apache.cxf.ws.security.sts.provider.STSException;
 import org.apache.cxf.ws.security.sts.provider.model.RequestSecurityTokenResponseCollectionType;
-import org.apache.cxf.ws.security.sts.provider.model.RequestSecurityTokenType;;
+import org.apache.cxf.ws.security.sts.provider.model.RequestSecurityTokenType;
 import org.apache.wss4j.common.WSS4JConstants;
 import org.apache.wss4j.common.principal.CustomTokenPrincipal;
 import org.w3c.dom.Element;
@@ -46,14 +45,16 @@ import java.io.StringWriter;
 import java.util.Collections;
 import java.util.Map;
 
-import static org.wso2.carbon.identity.sts.passive.utils.STSUtil.addSTSProperties;
-import static org.wso2.carbon.identity.sts.passive.utils.STSUtil.addService;
-import static org.wso2.carbon.identity.sts.passive.utils.STSUtil.addTokenProvider;
-import static org.wso2.carbon.identity.sts.passive.utils.STSUtil.changeNamespaces;
-import static org.wso2.carbon.identity.sts.passive.utils.STSUtil.createAppliesToElement;
-import static org.wso2.carbon.identity.sts.passive.utils.STSUtil.createSecondaryParameters;
-import static org.wso2.carbon.identity.sts.passive.utils.STSUtil.issueToken;
-import static org.wso2.carbon.identity.sts.passive.utils.STSUtil.setupMessageContext;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.addSTSProperties;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.addService;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.addTokenProvider;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.changeNamespaces;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.createAppliesToElement;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.createSecondaryParameters;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.getClaimURIs;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.getFormattedClaims;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.issueToken;
+import static org.wso2.carbon.identity.sts.passive.utils.RequestProcessorUtil.setupMessageContext;
 
 public class SigningRequestProcessor extends RequestProcessor {
 
@@ -66,6 +67,7 @@ public class SigningRequestProcessor extends RequestProcessor {
 
         ResponseToken responseToken;
         String tenantDomain = null;
+
         try {
             tenantDomain = request.getTenantDomain();
             int tenantId = IdentityPassiveSTSServiceComponent.getRealmService().getTenantManager()
@@ -77,19 +79,13 @@ public class SigningRequestProcessor extends RequestProcessor {
             }
 
             TokenIssueOperation issueOperation = new TokenIssueOperation();
-
-            addTokenProvider(issueOperation);
-
+            addTokenProvider(issueOperation, request);
             addService(issueOperation, request.getRealm());
-
             addSTSProperties(issueOperation, "localhost");
-
             // Set the ClaimsManager to the issue operation.
-            ClaimsManager claimsManager = new ClaimsManager();
-            ClaimsHandler claimsHandler = new CustomClaimsHandler();
-            claimsManager.setClaimHandlers(Collections.singletonList(claimsHandler));
-            issueOperation.setClaimsManager(claimsManager);
+            handleClaims(request, issueOperation);
 
+            // Obtain the type of token, whether it is a SAML1.1 or SAML2 token.
             String requestedTokenType = PassiveSTSUtil.extractTokenType(request);
 
             // Mock up an issue request.
@@ -105,22 +101,23 @@ public class SigningRequestProcessor extends RequestProcessor {
                 );
             }
             issueTokenRequest.getAny().add(tokenType);
-            Element secondaryParameters = createSecondaryParameters();
+            Element secondaryParameters = createSecondaryParameters(request);
             issueTokenRequest.getAny().add(secondaryParameters);
             issueTokenRequest.getAny().add(createAppliesToElement(request.getRealm()));
-
             Map<String, Object> msgCtx = setupMessageContext(request.getUserName());
 
+            // Make an issue token request.
             RequestSecurityTokenResponseCollectionType securityTokenResponse = issueToken(issueOperation, issueTokenRequest,
                     new CustomTokenPrincipal(request.getUserName()),
                     msgCtx);
 
+            // Convert the response into a JAXBElement.
             JAXBElement<RequestSecurityTokenResponseCollectionType> jaxbResponse =
                     QNameConstants.WS_TRUST_FACTORY.createRequestSecurityTokenResponseCollection(securityTokenResponse);
 
+            // Create XML Formatted Response.
             StringWriter sw = new StringWriter();
             try {
-                // Create XML Formatted Response.
                 JAXBContext jaxbContext = JAXBContext.newInstance(RequestSecurityTokenResponseCollectionType.class);
                 Marshaller jaxbMarshaller = jaxbContext.createMarshaller();
                 jaxbMarshaller.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT, Boolean.TRUE);
@@ -146,5 +143,23 @@ public class SigningRequestProcessor extends RequestProcessor {
         }
 
         return responseToken;
+    }
+
+    /**
+     * Handles the claim related logic. For example setting the known URIs.
+     *
+     * @param requestToken Request sent by the client to obtain the security token.
+     * @param issueOperation The issue operation which the claim manager should be set into.
+     */
+    private static void handleClaims(RequestToken requestToken, TokenIssueOperation issueOperation) {
+
+        CustomClaimsHandler customClaimsHandler = new CustomClaimsHandler();
+        CustomClaimsHandler.setKnownURIs(getClaimURIs(requestToken));
+        customClaimsHandler.setClaimsKeyValuePair(getFormattedClaims(requestToken));
+
+        ClaimsManager claimsManager = new ClaimsManager();
+        claimsManager.setClaimHandlers(Collections.singletonList(customClaimsHandler));
+
+        issueOperation.setClaimsManager(claimsManager);
     }
 }
